--- randomizer.js
@@ -1,3 +1,0 @@
-export const randomizer = num => Math.ceil(Math.random() * num);
-
-
